refactor: move Dialer from iroh-gossip to iroh-net
