+++ iroh-gossip/src/net/util.rs
@@ -1,18 +1,13 @@
 //! Utilities for iroh-gossip networking
 
-use std::{collections::HashMap, io, pin::Pin, time::Instant};
+use std::{io, pin::Pin, time::Instant};
 
-use anyhow::{anyhow, bail, ensure, Context, Result};
+use anyhow::{bail, ensure, Context, Result};
 use bytes::{Bytes, BytesMut};
-use futures::future::BoxFuture;
-use iroh_net::{key::PublicKey, MagicEndpoint, NodeAddr, NodeId};
 use tokio::{
     io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt},
-    task::JoinSet,
     time::{sleep_until, Sleep},
 };
-use tokio_util::sync::CancellationToken;
-use tracing::error;
 
 use crate::proto::util::TimerMap;
 
@@ -76,118 +71,6 @@ pub async fn read_lp(
         }
     }
     Ok(Some(buffer.split_to(size).freeze()))
-}
-
-/// Future for a pending dial operation
-pub type DialFuture = BoxFuture<'static, (PublicKey, anyhow::Result<quinn::Connection>)>;
-
-/// Dial nodes and maintain a queue of pending dials
-///
-/// This wraps a [`MagicEndpoint`], connects to nodes through the endpoint, stores
-/// the pending connect futures and emits finished connect results.
-///
-// TODO: Move to iroh-net
-#[derive(Debug)]
-pub struct Dialer {
-    endpoint: MagicEndpoint,
-    pending: JoinSet<(PublicKey, anyhow::Result<quinn::Connection>)>,
-    pending_dials: HashMap<PublicKey, CancellationToken>,
-}
-
-impl Dialer {
-    /// Create a new dialer for a [`MagicEndpoint`]
-    pub fn new(endpoint: MagicEndpoint) -> Self {
-        Self {
-            endpoint,
-            pending: Default::default(),
-            pending_dials: Default::default(),
-        }
-    }
-
-    /// Start to dial a node.
-    ///
-    /// Note that the node's addresses and/or derp url must be added to the endpoint's
-    /// addressbook for a dial to succeed, see [`MagicEndpoint::add_node_addr`].
-    pub fn queue_dial(&mut self, node_id: NodeId, alpn_protocol: &'static [u8]) {
-        if self.is_pending(&node_id) {
-            return;
-        }
-        let cancel = CancellationToken::new();
-        self.pending_dials.insert(node_id, cancel.clone());
-        let endpoint = self.endpoint.clone();
-        self.pending.spawn(async move {
-            let res = tokio::select! {
-                biased;
-                _ = cancel.cancelled() => Err(anyhow!("Cancelled")),
-                res = endpoint.connect(NodeAddr::new(node_id), alpn_protocol) => res
-            };
-            (node_id, res)
-        });
-    }
-
-    /// Abort a pending dial
-    pub fn abort_dial(&mut self, node_id: &NodeId) {
-        if let Some(cancel) = self.pending_dials.remove(node_id) {
-            cancel.cancel();
-        }
-    }
-
-    /// Check if a node is currently being dialed
-    pub fn is_pending(&self, node: &NodeId) -> bool {
-        self.pending_dials.contains_key(node)
-    }
-
-    /// Wait for the next dial operation to complete
-    pub async fn next_conn(&mut self) -> (PublicKey, anyhow::Result<quinn::Connection>) {
-        match self.pending_dials.is_empty() {
-            false => {
-                let (node_id, res) = loop {
-                    match self.pending.join_next().await {
-                        Some(Ok((node_id, res))) => {
-                            self.pending_dials.remove(&node_id);
-                            break (node_id, res);
-                        }
-                        Some(Err(e)) => {
-                            error!("next conn error: {:?}", e);
-                        }
-                        None => {
-                            error!("no more pending conns available");
-                            futures::future::pending().await
-                        }
-                    }
-                };
-
-                (node_id, res)
-            }
-            true => futures::future::pending().await,
-        }
-    }
-
-    /// Number of pending connections to be opened.
-    pub fn pending_count(&self) -> usize {
-        self.pending_dials.len()
-    }
-}
-
-impl futures::Stream for Dialer {
-    type Item = (PublicKey, anyhow::Result<quinn::Connection>);
-
-    fn poll_next(
-        mut self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Option<Self::Item>> {
-        match self.pending.poll_join_next(cx) {
-            std::task::Poll::Ready(Some(Ok((node_id, result)))) => {
-                self.pending_dials.remove(&node_id);
-                std::task::Poll::Ready(Some((node_id, result)))
-            }
-            std::task::Poll::Ready(Some(Err(e))) => {
-                error!("dialer error: {:?}", e);
-                std::task::Poll::Pending
-            }
-            _ => std::task::Poll::Pending,
-        }
-    }
 }
 
 /// A [`TimerMap`] with an async method to wait for the next timer expiration.
